extract team form from sheet
